--- src/components/layout/AuthLayout/index.ts
@@ -1,1 +1,0 @@
-export { default as AuthLayout } from "./AuthLayout";
