+++ src/router/designer/index.tsx
@@ -8,9 +8,8 @@ import DesignerSignUpDetailComplete from "../../pages/designer/signup-detail-com
 import DesignerMyBadgesPage from "../../pages/designer/mypage/badges";
 import DesignerHome from "../../pages/designer/home";
 import { UserLayout } from "../../components/layout/UserLayout";
-import { AuthLayout } from "../../components/layout/AuthLayout";
+import { Layout } from "../../components/layout/Layout";
 import Quote from "../../pages/designer/quote";
-import { Layout } from "../../components";
 import theme from "../../style/theme";
 
 export const designerPaths = [
@@ -34,7 +33,7 @@ export const designerPaths = [
         element: <DesignerSignUpDetailComplete />,
       },
     ],
-    element: <AuthLayout />,
+    element: <Layout />,
   },
   {
     path: "",
